Refactored Code + used localappdata
