+++ auratext/Core/WelcomeScreen.py
@@ -1,22 +1,34 @@
+from __future__ import annotations
+from typing import TYPE_CHECKING
+
 from PyQt6.QtCore import Qt
 from PyQt6.QtGui import QFont, QPixmap
 from PyQt6.QtWidgets import QPushButton, QWidget, QVBoxLayout, QLabel
 
+
+if TYPE_CHECKING:
+    from auratext.Core.window import Window
+
+
 class WelcomeWidget(QWidget):
-    def __init__(self, main):
+    def __init__(self, window: Window):
         super().__init__()
 
         layout = QVBoxLayout()
-        #layout.addStretch(102)
+        # layout.addStretch(102)
 
-        label2 = QLabel(self)
-        pixmap = QPixmap('cat.jpg')
+        label2 = QLabel()
+        label2.setAlignment(Qt.AlignmentFlag.AlignHCenter)
+        pixmap = QPixmap(f"{window.local_app_data}/icons/splash_morning.png")
         label2.setPixmap(pixmap)
         self.resize(pixmap.width(), pixmap.height())
 
         label = QLabel("Welcome to Aura Text")
-        label3 = QLabel("Aura Text is a versatile and powerful text editor powered by QScintilla that provides all the " + "\n" +
-                        "necessary tools for developers. It is build using PyQt6 and Python.")
+        label3 = QLabel(
+            "Aura Text is a versatile and powerful text editor powered by QScintilla that provides all the "
+            + "\n"
+            + "necessary tools for developers. It is build using PyQt6 and Python."
+        )
         label1 = QLabel("Please open a folder or create a new file to start editing.")
 
         label.setAlignment(Qt.AlignmentFlag.AlignCenter)
@@ -29,10 +41,10 @@ class WelcomeWidget(QWidget):
         label3.setFont(QFont("Consolas", 12))
 
         button = QPushButton("Open Folder")
-        button.clicked.connect(main.open_project)
+        button.clicked.connect(window.open_project)
 
         button1 = QPushButton("Clone from Git")
-        button1.clicked.connect(main.gitClone)
+        button1.clicked.connect(window.gitClone)
 
         layout.addWidget(label2)
         layout.addWidget(label)
