--- auratext/Data/CPath_Project.txt
@@ -1,1 +1,0 @@
-C:/Users/rohan/PycharmProjects/AuraText PyQt6
