+++ auratext/Data/config.json
@@ -1,1 +1,3 @@
-{"intend_length": 4}
+{
+    "intend_length": 4
+}
